requirements.txt
beautifulsoup4==4.8.0
bootstrap4==0.1.0
certifi==2018.8.24
chardet==3.0.4
cssselect==1.1.0
Django==2.2.5
django-appconf==1.0.3
django-background-task==0.1.8
django-background-tasks==1.2.0
django-bootstrap==0.2.4
django-bootstrap-static==4.0.0
django-bootstrap4==1.0.1
django-compat==1.0.15
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-extensions==2.2.1
django-fontawesome==1.0
django-jquery==3.1.0
django-ranged-response==0.2.0
django-render-partial==0.4
django-rest-framework==0.1.0
django-select2==7.1.1
django-simple-captcha==0.5.12
django-treemenus==0.9.2
djangorestframework==3.10.3
googlemaps==3.1.3
idna==2.8
lxml==4.4.1
#Pillow==6.1.0
psycopg2-binary==2.8.3
pyquery==1.4.0
python-google-places==1.4.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
urllib3==1.25.5
selenium==3.141.0